Fix search bag
When we search by RENT we receive RENT and NON_RENT results
